86czyq5gc- fix up of use case to use types
